unresolved symbols repaired, data transfer failed

+++ cuda/core/GpuBasics.cu
@@ -3,7 +3,7 @@
 
 #include <iostream>
 
-void copyH2D(double** dst, double** src, size_t N) {
+void copyH2D(double** dst, double** src, int N) {
 	cudaError_t err;
 	err =
 		cudaMemcpy(dst, src, N * sizeof(double*), cudaMemcpyHostToDevice);
